backend/requirements.txt
fastapi
uvicorn[standard]
supabase
google-genai
pydantic
python-dotenv
